Fix bug in no_results
The code was interacting with `no_results_html` as if it was a jQuery
object. When we upgraded to v1.7.0 the code changed from constructing
a jQuery object from an HTML string to using the `get_no_results_html`
method. This method only returns a string of HTML and hence we need to
wrap it in a call to jQuery to get the `.find` call to work.